+++ hutool-extra/src/main/java/cn/hutool/extra/pinyin/engine/tinypinyin/TinyPinyinEngine.java
@@ -1,5 +1,7 @@
 package cn.hutool.extra.pinyin.engine.tinypinyin;
 
+import cn.hutool.core.lang.Opt;
+import cn.hutool.core.util.StrUtil;
 import cn.hutool.extra.pinyin.PinyinEngine;
 import com.github.promeg.pinyinhelper.Pinyin;
 
@@ -51,7 +53,8 @@ public class TinyPinyinEngine implements PinyinEngine {
 
 	@Override
 	public String getPinyin(String str, String separator) {
-		return Pinyin.toPinyin(str, separator).toLowerCase();
+		final String pinyin = Pinyin.toPinyin(str, separator);
+		return StrUtil.isEmpty(pinyin) ? pinyin : pinyin.toLowerCase();
 	}
 
 }